# .agents/skills/payluk-api/references/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payluk ThirdParty API",
    "version": "1.0.0",
    "description": "Merchant / Business API for the Payluk escrow and payments platform.\n\nAll routes are mounted under `/v1` and authenticated with a secret key. Every response uses the standard envelope `{ status, message, data }`.\n\n**Rate limit.** Each secret key may make **10 requests per minute** across all `/v1` routes. Beyond that the API answers `429` with `{ \"status\": 429, \"message\": \"Too many request\" }`. Every response carries `RateLimit` and `RateLimit-Policy` headers (IETF draft-7) so a client can back off before being cut off.\n\n**IP allowlist.** On production, if you have added IP addresses to the allowlist in your dashboard, requests from any other address are refused with `403 Unauthorized IP address`. An empty allowlist admits every address. Staging never checks the caller's IP.\n\n**Merchant super-admin scope.** Escrow, milestone-escrow and category routes accept any approved business key. Every other route (customers, wallet, merchant account, confirming and disputing, additional fee, vault, payments, cards, countries, crypto whitelist) requires the key to belong to a merchant super-admin account and answers `401 Access denied` otherwise.",
    "contact": {
      "name": "Payluk Developer Support",
      "email": "support@payluk.ng",
      "url": "https://payluk.ng"
    }
  },
  "servers": [
    {
      "url": "https://staging.api.payluk.ng",
...
    },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/payment/verify": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Verify payment",
        "operationId": "verifyPayment",
        "description": "**Submits a previously created payment intent for processing.** Payment uses a deliberate two-step flow for security: [Create payment intent](/api-reference/payments/create-payment-intent) only stages the transaction (no money moves), and this endpoint is what actually executes it: debiting the wallet and carrying out the withdrawal, crypto transfer, card charge, or wallet transfer.\n\nPass the single `reference` returned when the intent was created. If the transaction requires confirmation (e.g. an SMS OTP / transaction PIN), include it as `otp`. Requires the `customer-id` header.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CustomerIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reference"
                ],
                "properties": {
                  "reference": {
                    "type": "string",
                    "description": "The reference of the payment intent to submit for processing.",
                    "example": "99999533334"
                  },
                  "otp": {
                    "type": "string",
                    "description": "Confirmation OTP / transaction PIN. Required only when the transaction needs confirmation.",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment intent submitted and processed.",
            "content": {
              "application/json": {
                "example": {
                  "status": 200,
                  "message": "Payment intent verified successfully",
                  "data": {
                    "id": "6a3cea83734e96016b227460",
                    "amount": 1000,
                    "reference": "t53gtryhtyut",
                    "fee": 0,
                    "transactionType": "wallet_transfer",
                    "currency": "NGN",
                    "transferDetails": null,
                    "cardId": null,
                    "walletDetails": {
                      "phone": "09022334422",
                      "name": "King David",
                      "narration": "test"
                    },
                    "blockchainDetails": null,
                    "withdrawalDetails": null,
                    "escrowDetails": null,
                    "metadata": null,
                    "status": "success",
                    "creditType": "debit",
                    "createdAt": "2026-06-25T08:44:51.817Z",
                    "updatedAt": "2026-06-25T08:46:39.479Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your secret key as a Bearer token. The key prefix selects the environment: `sk_test_...` (staging) or `sk_live_...` (production); a key on the wrong host is refused with `403 Unauthorized Access`. Each key is limited to 10 requests per minute (`429` beyond that) and, on production, to the IP addresses on your dashboard allowlist when one is configured."
      }
    },
    "parameters": {
...
      "CustomerIdHeader": {
        "name": "customer-id",
        "in": "header",
        "required": true,
        "description": "The merchant customer this request acts on behalf of.",
        "schema": {
          "type": "string"
        },
        "example": "665f1b2c9a1e4d0012ab3c01"
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "status": 400,
              "message": "amount is required",
              "data": {}
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded: more than 10 requests in one minute from this secret key. Read the `RateLimit` header and retry after the window resets.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "status": 429,
              "message": "Too many request"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Standard error envelope.",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Action not allowed"
          },
          "data": {
            "type": "object",
            "example": {}
          }
        }
      },
...
    }
  }
}